Moved externally accessible vars to config module. Fixes in `pad`
Now that theano_shim is a package, the from .core import * statement in
__init__.py did import the variables, but since they were in the __init__.py's
scope, they were separate from the variable instance in core.py. So changes in
core.py were not reflected in the outside interface. Having everyone access the
same variable with config.<var> solves that problem.
Variables moved:
 - use_theano
 - inf
 - theano_updates
 - RandomStreams
 - lib (Deprecated)