Fixed Numeric data type (moved to config.py)
Undoes the earlier integration today of the is_numeric function, so that the
Numeric type does what it's actually supposed to -- namely to provide type
hints. This only works with Python >=3.5, so a version guard is also added.

I don't use type hints at the moment, so Giacomo should test whether this works
for his use case.