Rewrite graph.GraphExpression properly.
Old version attempted some monkey patching to avoid having to learn how 
some of the deep Theano internals work. It never worked, and having now 
learned enough of the internals to do this properly, I rewrote this 
function.
The new implementation subclasses the Elementwise operation, to create 
an Elementwise{identity} which returns a Variable node of the desired 
type.